requirements.txt
flask
flask-cors
gunicorn
langchain
langchain-openai
langchain-ollama
langchain-pinecone
langgraph
openai
pinecone-client
python-dotenv
tiktoken
